move owner from ship to object
we want to remove ships when their life is reduced to zero,
however, the rows has to be kept in the database as they
are referenced as foreign keys in some other tables
(namely combat reports and trade reports).

to make removing ships possible, those foreing keys now point
to corresponding objects rather than actual ships,
therefore it is possible to remove the ship without removing the object,
effectively leaving an unusable "wreck" behind the ship.

to make the reports meaningfull, it is desirable to know owners of the former ships,
therefore the owner (formerly player) field has been moved to the object from the ship.

planets (and waypoints) have NULL owner, while wrecks (and ships) have their owner preserved.

--

Furthermore, this commit fixes that it was not possible to construct non-shipyard ships,
and that any ship could construct new ships.